Added New LAN Game Window
Added a window in Main Menu for new LAN game.  It's all hooked up for keyboard UI navigation, and it has the appropriate choices for LAN host or dedicated server, and choosing how many players are on the local connection.  The next step would be for this to go to a lobby screen, but that doesn't exist yet.  It doesn't do anything network-related yet either.